Aviso creado y Seguridad al 100%, ESTÁ VEZ SÍ

+++ David_Falla_Urtiaga_Proyecto_Guarderia/src/main/java/com/salesianostriana/dam/proyecto_guarderia/controlador/CursoControlador.java
@@ -13,6 +13,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 
 import com.salesianostriana.dam.proyecto_guarderia.modelo.Curso;
 import com.salesianostriana.dam.proyecto_guarderia.servicio.CursoServicio;
+import com.salesianostriana.dam.proyecto_guarderia.servicio.ProfesorServicio;
 
 @Controller
 @RequestMapping("/admin/cursos")
@@ -21,9 +22,10 @@ public class CursoControlador {
 	@Autowired
 	private CursoServicio servicio;
 	
+	
 	//MOSTRAR LA LISTA DE CURSOS -----------------------------------------------------------------------------
 	
-	@GetMapping("")
+	@GetMapping("/")
 	public String mostrarCursos(Model model) {
 			
 		model.addAttribute("listaCursos", servicio.findAll());
@@ -52,7 +54,7 @@ public class CursoControlador {
 			
 			servicio.save(curso);		
 			
-		return "redirect:/admin/cursos";
+		return "redirect:/admin/cursos/";
 	} 
 
 	
@@ -61,16 +63,16 @@ public class CursoControlador {
 	@GetMapping("/editarCurso/{id}")
 	public String mostrarFormularioCursosEditar(@PathVariable("id") long id, Model model) {
 			
-		Optional<Curso> cursosAEditar = servicio.findById(id);
+		Optional<Curso> cursoAEditar = servicio.findById(id);
 			
-		if(cursosAEditar.isPresent()) {
+		if(cursoAEditar.isPresent()) {
 				
-			model.addAttribute("curso", cursosAEditar.get());
+			model.addAttribute("curso", cursoAEditar.get());
 			return "admin/agregarEditarCursosAdmin";
 				
 		} else {
 				
-			return "redirect:/admin/cursos";
+			return "redirect:/admin/cursos/";
 		}
 			
 	}
@@ -84,7 +86,7 @@ public class CursoControlador {
 			
 		servicio.save(curso);
 			
-		return "redirect:/admin/cursos";	
+		return "redirect:/admin/cursos/";	
 	}
 		
 		
@@ -92,11 +94,20 @@ public class CursoControlador {
 	//BORRA EL CURSO ELEGIDO POR ID ----------------------------------------------------------------------
 
 	@GetMapping("/borrarCurso/{id}")
-	public String borrarCurso(@PathVariable("id") long id) {
+	public String borrarCurso(@PathVariable("id") long id, Model model) {
+
+		Optional<Curso> cursoABorrar = servicio.findById(id);
+		
+		if(cursoABorrar.isPresent()) {
+		
+			if(servicio.contarProfesoresDeUnCurso(id) == 0) {
+				servicio.deleteById(id);
+			}else {
+				return "redirect:/admin/cursos/?error=true";
+			}
+		}
 			
-		servicio.deleteById(id);
-			
-		return "redirect:/admin/cursos";
+		return "redirect:/admin/cursos/";
 	}
